move tests. better tests
